requirements.txt
async-timeout==3.0.1
chardet==3.0.4
discord-py==1.3.3
PyNaCl==1.3.0
dnspython==1.16.0
